Refactor: extract EntryDetailWidget
